requirements/base.txt
# API framework
fastapi==0.104.1
uvicorn==0.24.0
mangum==0.17.0
pydantic==2.4.2
pydantic-settings==2.0.3
email-validator==2.1.0.post1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AWS
boto3==1.28.64
boto3-stubs[dynamodb]==1.28.64

# Logging
python-json-logger==2.0.7

# Date/time utilities
pytz==2023.3

# Development utilities
python-dotenv==1.0.0